+++ src/main/java/store/service/strategy/AddFreePromotionStrategy.java
@@ -5,7 +5,7 @@ import store.domain.vo.Order;
 import store.domain.Promotion;
 import store.domain.vo.PromotionQueryResult;
 import store.domain.vo.*;
-import store.service.dto.PromotionCommandResponse;
+import store.service.dto.response.PromotionCommandResponse;
 
 /**
  * 적용 가능한 프로모션 제품을 주문 내역에 추가하는 전략
