+++ src/main/java/store/service/strategy/ImmutableOrderPromotionStrategy.java
@@ -5,7 +5,7 @@ import store.domain.vo.Order;
 import store.domain.Promotion;
 import store.domain.vo.PromotionQueryResult;
 import store.domain.vo.*;
-import store.service.dto.PromotionCommandResponse;
+import store.service.dto.response.PromotionCommandResponse;
 
 /**
  * 기존의 주문 내역을 변경하지 않고 프로모션을 적용하는 전략
