+++ src/main/java/store/service/strategy/PromotionStrategy.java
@@ -3,7 +3,7 @@ package store.service.strategy;
 import store.domain.vo.Inventory;
 import store.domain.vo.Order;
 import store.domain.Promotion;
-import store.service.dto.PromotionCommandResponse;
+import store.service.dto.response.PromotionCommandResponse;
 
 public interface PromotionStrategy {
     /**
